refactor(types): type 관련 파일 분리 및 적용
각각의 컴포넌트에서 정의하던 type들을 store폴더 내 저장
