+++ src/components/Modal/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { createPortal } from 'react-dom';
 import Store from '../../store';
+import { Restaurant } from '../../store/type';
 import RestaurantItem from '../RestaurantItem';
 import styles from './Modal.module.css';
-import type { Restaurant } from '../RestaurantItem/type';
 
 class Modal extends React.PureComponent {
 	private static findRestaurantById = (restaurantList?: Restaurant[], id?: string) =>
